Some changes in the structure
- LoggingConfiguration -> gLogging : a kind of logging wrapper
- Backends = wrapper of handler + format
