+++ FrameworkSystem/private/standardLogging/Backend/FileBackend.py
@@ -1,4 +1,5 @@
 import logging
+from os import getpid
 
 from DIRAC.FrameworkSystem.private.standardLogging.Backend.Backend import Backend
 from DIRAC.FrameworkSystem.private.standardLogging.Formatter.BaseFormatter import BaseFormatter
@@ -12,6 +13,10 @@ class FileBackend(Backend):
 
   def __init__(self):
     super(FileBackend, self).__init__(None, BaseFormatter())
+    self.fileName = 'Dirac-log_%s.log' % getpid()
 
   def setParameters(self, parameters):
-    self.handler = logging.FileHandler(parameters['FileName'])
+    self.fileName = parameters['FileName']
+
+  def configureHandler(self):
+    self.handler = logging.FileHandler(self.fileName)
